Add files via upload
This calculator is a part of my web development internship project. It is built using HTML, CSS, and JavaScript with a clean and intuitive interface. The calculator performs basic arithmetic operations like addition, subtraction, multiplication, and division. Designed to be responsive and easy to use, it showcases my understanding of DOM manipulation and event handling in JavaScript. This project helped me improve my front-end development skills and gave me hands-on experience in building functional UI components.